Fix another IRQ clash; revert to invalid packet sizes
Also remove an unused transmit buffer...

For some reason just halving the buffer size and doubling the frame interval isn't working for input to the Teensy, so for the present we'll test without. Very odd - it seems to do a burst of transfers at 500µs intervals, but then stop, so on average there's only a transfer every millisecond. The data get scrambled, too.